Quest - The Hunter's Path (Teldrassil):
"Dazalar in Dolanaar has nothing to say when he is reported to.
His line is missing or was never meant to be there in the first place.
Never the less, this should be fixed."

I am sure this was forgotten by the Oregon team.
Now Dazalar is no longer muted when the player reports to him.

--HG--
branch : pool