Resolved issue #456 - Cat Form
Some spells like Cat Form didn't work correctly if vmap outdoor checks were enabled.
Cat form was unable to even use in indoors because its auras were cancelling it.
The aura was Feral Swiftness. Even though it could be hackfixed, it still wouldn't work as expected.
With hackfix it works like this:
    - ferel swiftness works also in indoors (if casted indoors)
    - feral swiftness will be cancelled when one comes from outdoors to indoors
The feral swiftness is required to work only outside.

With this fix it works like it should have:
    - if cat form is casted indoors, feral swiftness is not applied
    - if you casted cat form indoors and come outdoors, FS will be applied automatically
    - if you casted cat form outdoors the FS buff is applied automatically,
      then if you enter indoors, it will be cancelled, and if you return outdoors,
      it will be applied again.
This fix is not a hackfix but (I think) a good way to do this also for other spells but I can't recall any of them.

--HG--
branch : devel