Close #692
I changed the error message to be more verbose, but unfortunately this cannot be fixed because its probably
very specific to your configuration. That error simply means that your petition is missing from db. If this
issue wasn't caused by manipulating with db, this can be eventually fixed. I'll need steps to reproduce though.