Fix issue #718
Sorry for the delay... I'm quite busy right now.
The commit that caused this commit was on the good way,
but still made some troubles around with aura stacking.

Auras should stack even if they are not stackable but casted
from different enchanted items/weapons. So for now, let's
keep the mongoose stacking hackfixed. The others spells
will follow when we'll know how to better solve that issue.