Add minimum range to linked GameObject search
Currently linked GameObjects are searched for with search range equal to the range of the spell that the linked GameObject will cast. This causes some linked GameObject traps to not work at all, like Musty Tome (176151) for example, because the spell has range 0 (but it is definitely intended to be cast).