[Core/Spells] Fix exploit/bug with Nether-wraith beacon
In quest 10832 you are given an item that summons two creatures to attack you when used in Netherstorm. Previously this item could be spammed to summon hundreds of these nether wraiths because the wraiths were summoned when the item was clicked, not when the item spell finishes casting. This is still somewhat exploitable after this change because the item has no cooldown, which might be blizzlike based on http://www.wowhead.com/item=31742/nether-wraith-beacon, so maybe more changes needs to be done.