Core/Aura: Don't deal damage to neutral guards if not at war
unfortunately, this still shows in combat log that the target is
afflicted by the aura. i am not sure if this is correct behaviour or not
(most likely not). so while the underlying issue is solved there is a
better way to do it that involves the aura not even applying to the
guard

Updates #1559